hjkl

+++ libft/ft_atoi.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
 	int			nv;
 	long int	nb;
